Network scan cancel button fix
