+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pynetwizard"
-version = "0.5.0"
+version = "0.5.1"
 description = "Graphical network device configurator"
 readme = "README.md"
 requires-python = ">=3.13"
